It's a little messy, but it works.
I used Image::ExifTool; it may not be the best way to identify files as JPGs and determine their dimensions, but I already had it installed. A bird in the hand, etc...

The file comparison starts with file size: if they're not the same size then they do not match.