It seems to contain information directly related to these images from C:\Users\pete\AppData\Local\Packages\Microsoft.Windows.ContentDeliveryManager_cw5n1h2txyewy\LocalState\Assets
2ed... is the Porcupine Mountains image
f71... is the Faroe Islands image

There would also have been a portrait version in the Assets folder, but we are not concerned with those at this point.

What we now need to determine is whether the same file is updated for each new Spotlight image delivery, although at this point it seems likely that for each new delivery, a new numbered folder is created under C:\Users\pete\AppData\Local\Packages\Microsoft.Windows.ContentDeliveryManager_cw5n1h2txyewy\LocalState\ContentManagementSDK\Creatives.

The data file itself appears to be JSON, so perhaps we can parse it to get the information we require. The important question is how we determine which JSON file relates to which Spotlight Images.