Controversy!
The recommended method for implementing a singleton is to return an unexported type from an Exported func -- but this triggers lint warning #210. There seems to be no clear decision on the interwebs (at least that I could find) as to the best way to handle this.

I think for my own purposes I can get away with simply Exporting the paths struct--but I'll need to bear this in mind for the future.