Given the effort I just had forcing my code to read a UTF-16 JSON file, it seems a bit naive (even rude) that my code then strips out unicode characters from description and copyright messages, and hence from the target filename.
I need to fix this. I've lived too long in my nice cosy ASCII-world bubble; it's time my code grew up and started being Unicode-friendly! If the OS can support the characters in a filename, I should not be dumping them out.

Also? This means it's time to stop using editors which don't properly handle them when in a file. Sadly, since SciTE mangles "Kjölur" into "KjÃ¶lur"--one of the main reasons I had so much trouble identifying the UTF-16 issue in the first place was because I had opened and saved my test metadata file in SciTE)--this means it's time to stop using it for "real" coding.