Added newFilename() to handle the formatting of the target filename; it now takes the more friendly approach of stripping out disallowed characters rather than trying to pass through allowed characters.
Also, version = 1.2? In Perl there is a lot of (conflicting) discussion around how best to handle module versioning. There is https://blog.golang.org/versioning-proposal to consider, but it seems it's still early days for Go. That said, I'm really not sure how best to handle *release* versioning. It seems to me it would be nice to have the version appear in an output string, and I've been using semantic versioning in my GitHub releases. Is there a preferred way to sync the two up? For now I'll call this 1.2 because I know that will be my next official release, but -- I dunno. Maybe there is a better way?