decouple ini file
When I first built this code I wasn't fully up to speed on the best way to write
Go code. As a result, despite there apparently being nebulous plans for a whole
"wallpaper toolset", several of the packages (which should probably be reusable)
have "UpdateWallpaper" details hardcoded within them. We need to decouple the
packages from the cmd logic.